New functions, bug fixes, and tweaks to Flash.cs
 - Add delay between retries to several operations. It appears that the
flashing code gets overwhelmed and the TCPIP stack starts dropping data
packets. Having a delay between retry attempts seems to help the stack
catch up.

 - Add extra read in autobaud detection to catch a second OK that
flasher sends. It appears to send OK to both "synchronized" and the
clock rate, I think it is only supposed to send OK to the clock rate
but that is not clear. Having an extra read at the end of the function
to helps pulls the OK out of the receive buffer so that it does not
trip up subsequent commands.

 - Add retry support to WriteToRAM.

 - Add CopyRAMtoSector() to handle the logic to copy a large chunk of
RAM to a sector.

 - Fix bug in calculating for FPS checksums for the first sector being
flashed. Checksum2End_Address was being used where Checksum2_Address
should have been used.